deleted DFSVisualizer and BFSVisualizer. created AlgorithmVisualizer class. fixed DFS and BFS. added IAlgorithm interface
