+++ src/main/java/gui/IAlgorithmVisualizer.java
@@ -1,0 +1,9 @@
+package gui;
+
+import java.awt.*;
+
+public interface IAlgorithmVisualizer {
+    void paint(Graphics g, int centerX, int centerY);
+    boolean isFinished();
+    void start();
+}
