+++ src/main/java/gui/MainApplicationFrame.java
@@ -9,7 +9,7 @@ import javax.swing.*;
 class MainApplicationFrame extends JFrame {
     private GifCreator gifCreator;
     private Visualizer visualizer;
-    MainApplicationFrame(Graph graph, AlgorithmVisualizer algorithmVisualizer, String gifname) {
+    MainApplicationFrame(Graph graph, IAlgorithmVisualizer algorithmVisualizer, String gifname) {
         setLocationRelativeTo(null);
         setPreferredSize(new Dimension(750,600));
         setDefaultCloseOperation(JFrame.DISPOSE_ON_CLOSE);
