+++ src/main/java/tests/TestDFS.java
@@ -15,6 +15,7 @@ import static org.junit.Assert.*;
 public class TestDFS {
     private boolean[][] adjacencyMatrix1;
     private boolean[][] adjacencyMatrix2;
+    private DFS dfs = new DFS();
 
     @Before
     public void setUp(){
@@ -32,7 +33,7 @@ public class TestDFS {
         expectedEdges.add(new Edge(3, 4));
         expectedEdges.add(new Edge(0, 1));
         expectedEdges.add(new Edge(1, 3));
-        ArrayList<Edge> actualEdges = DFS.DFS(graph, 0);
+        ArrayList<Edge> actualEdges = dfs.invoke(graph, 0);
         for (int i = 0; i < actualEdges.size(); i++) {
             assertEquals(expectedEdges.get(i), actualEdges.get(i));
         }
@@ -56,7 +57,7 @@ public class TestDFS {
         expectedEdges.add(new Edge(5, 4));
         expectedEdges.add(new Edge(4, 1));
         expectedEdges.add(new Edge(1, 0));
-        ArrayList<Edge> actualEdges = DFS.DFS(graph, 0);
+        ArrayList<Edge> actualEdges = dfs.invoke(graph, 0);
         for (int i = 0; i < actualEdges.size(); i++) {
             assertEquals(expectedEdges.get(i), actualEdges.get(i));
         }
